requirements.txt
Flask==2.1.0
requests==2.28.1
mysql-connector-python==8.0.31
matplotlib==3.5.1
